Refactor: 전역 스토어 및 커스텀 훅 기반 리팩토링
